bug fixes in RL wrt conversion of counter to rates

+++ server/src/main/java/com/flipkart/varadhi/qos/entity/SuppressionFactor.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class SuppressionFactor {
-    Float throughputFactor;
-    Float qpsFactor;
+    Double throughputFactor;
+    Double qpsFactor;
 }
